Refactored root node stuff back to SideBar. Added rootItem prop.
