+++ src/components/NavPages.tsx
@@ -1,8 +1,5 @@
-import { useLocation } from "@gatsbyjs/reach-router"
-import { graphql, Link, useStaticQuery } from "gatsby"
-import { useContext } from "react"
-import { PageContext } from "../contexts/PageContext"
-import { NavigationItem, StrapiNavigation } from "../types/strapi"
+import { Link } from "gatsby"
+import { NavigationItem } from "../types/strapi"
 
 type ItemProps = {
   item: NavigationItem
@@ -37,51 +34,11 @@ const Item: React.FC<ItemProps> = ({ item: { title, path, items } }) => {
   )
 }
 
-type NavQuery = {
-  allStrapiPublicNavigation: {
-    nodes: StrapiNavigation[]
-  }
+type Props = {
+  rootItem: NavigationItem | undefined
 }
 
-const NavPages = () => {
-  const { locale } = useContext(PageContext)
-
-  // get the nav structure in the current locale
-  const navigations = useStaticQuery(graphql`
-    query {
-      allStrapiPublicNavigation {
-        nodes {
-          locale
-          items {
-            title
-            path
-            items {
-              title
-              path
-              items {
-                title
-                path
-              }
-            }
-          }
-        }
-      }
-    }
-  `) as NavQuery
-  const nav = navigations.allStrapiPublicNavigation.nodes.find(
-    (n) => n.locale === locale
-  )
-
-  // use browser location to figure out which nav item to show in the sidebar
-  const { pathname } = useLocation()
-  const pathParts = pathname
-    .split("/")
-    .slice(1) // remove empty part before /
-    .filter((part, index) => index !== 0 || part !== locale) // remove locale name as first part
-  const rootItem = nav?.items?.find((item) =>
-    item.path.startsWith(`/${pathParts[0]}`)
-  )
-
+const NavPages: React.FC<Props> = ({ rootItem }) => {
   return (
     <nav className="flex flex-col min-h-full bg-gray-darkest text-white pl-10 pr-5">
       {rootItem?.items?.map((item) => (
